[Refactor][Quantization] Reorganize quantization module: group configs and methods into subdirectories
- Move 3 config files into configs/ subdirectory
- Group 16 method files into 5 subdirectories: w8a8/, w4a8/, wna16/, w4a4/, kv_cache/
- Merge fp8/ DS-specific classes into parent files
- Merge w4a16/ and w8a16/ into unified wna16/
- Rename kv/ to kv_cache/ for clarity
- Extract shared is_fused_moe_layer() to utils.py
- Extract shared solve_kronecker_decompose() to w4a4_flatquant.py
- Remove unused self.ep_group attributes and get_ep_group imports
- Unify external imports to use short paths from methods/__init__.py
- Fix test patch paths and formatting issues


# Conflicts:
#	tests/ut/quantization/methods/test_w8a8_mxfp8.py

# Conflicts:
#	vllm_ascend/quantization/configs/fp8_config.py
#	vllm_ascend/quantization/methods/fp8.py

# Conflicts:
#	vllm_ascend/quantization/__init__.py
#	vllm_ascend/quantization/configs/fp8_config.py
#	vllm_ascend/quantization/methods/__init__.py
#	vllm_ascend/quantization/utils.py
