+++ tests/ut/quantization/methods/a2/test_w4a4_flatquant.py
@@ -5,7 +5,7 @@ import torch
 import torch.nn as nn
 
 from tests.ut.quantization.conftest_quantization import create_linear_layer
-from vllm_ascend.quantization.methods.w4a4_flatquant import (
+from vllm_ascend.quantization.methods.w4a4.w4a4_flatquant import (
     KRONECKER_QUANT_MAX_BATCH_SIZE,
     AscendW4A4FlatQuantDynamicLinearMethod,
     batched_kronecker_quant,
@@ -38,7 +38,7 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         self.assertEqual(get_decompose_dim(100), (10, 10))
         self.assertEqual(get_decompose_dim(99), (9, 11))
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.torch_npu")
     def test_pack_int4_weights_npu_success(self, mock_torch_npu):
         """
         Tests weight packing using the mocked NPU kernel.
@@ -54,7 +54,7 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         mock_torch_npu.npu_convert_weight_to_int4pack.assert_called_once()
         self.assertTrue(torch.equal(result, mock_packed_tensor))
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.torch_npu")
     def test_large_batch_multiple_calls(self, mock_npu):
         batch_size = 50000
         x = torch.randn(batch_size, 24, 32)
@@ -76,7 +76,7 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         batched_kronecker_quant(x, left_trans, right_trans, 0.95)
         self.assertEqual(mock_npu.npu_kronecker_quant.call_count, num_chunks)
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.torch_npu")
     def test_exact_max_batch_size(self, mock_npu):
         batch_size = KRONECKER_QUANT_MAX_BATCH_SIZE
         x = torch.randn(batch_size, 24, 32)
@@ -146,7 +146,7 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         x = torch.randn(batch_size, self.input_size, dtype=self.params_dtype)
         return layer, x, m, n
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.torch_npu")
     def test_apply_small_batch(self, mock_torch_npu):
         """Tests the apply method with a batch size smaller than MAX_BATCH_SIZE."""
         batch_size = 128
@@ -163,8 +163,8 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         self.assertTrue(torch.allclose(output, mock_output + bias.to(self.params_dtype)))
         self.assertEqual(output.shape, (batch_size, self.output_size))
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.KRONECKER_QUANT_MAX_BATCH_SIZE", 10)
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.KRONECKER_QUANT_MAX_BATCH_SIZE", 10)
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.torch_npu")
     def test_apply_large_batch(self, mock_torch_npu):
         """Tests the apply method with a batch size larger than MAX_BATCH_SIZE."""
         batch_size = 25
@@ -192,7 +192,7 @@ class TestW4A4FlatQuantDynamic(unittest.TestCase):
         with self.assertRaisesRegex(ValueError, "FlatQuant transform matrices dimension mismatch"):
             self.method.apply(layer, x)
 
-    @patch("vllm_ascend.quantization.methods.w4a4_flatquant.pack_int4_weights")
+    @patch("vllm_ascend.quantization.methods.w4a4.w4a4_flatquant.pack_int4_weights")
     def test_process_weights_after_loading(self, mock_pack_weights):
         """Tests weight processing after loading, without transpose."""
         layer = nn.Module()
