+++ tests/ut/quantization/methods/test_moe_logical_experts.py
@@ -4,13 +4,13 @@ import pytest
 
 from vllm_ascend.ops.fused_moe.moe_utils import get_moe_num_logical_experts
 from vllm_ascend.quantization.methods.base import AscendLinearScheme, AscendMoEScheme
-from vllm_ascend.quantization.methods.w4a4_mxfp4 import AscendW4A4MXFP4DynamicFusedMoEMethod
-from vllm_ascend.quantization.methods.w4a8 import AscendW4A8DynamicFusedMoEMethod
-from vllm_ascend.quantization.methods.w4a8_mxfp4 import AscendW4A8MXFPDynamicFusedMoEMethod
-from vllm_ascend.quantization.methods.w4a16 import AscendW4A16FusedMoEMethod
-from vllm_ascend.quantization.methods.w4a16_mxfp4 import AscendW4A16MXFP4FusedMoEMethod
-from vllm_ascend.quantization.methods.w8a8_dynamic import AscendW8A8DynamicFusedMoEMethod
-from vllm_ascend.quantization.methods.w8a8_mxfp8 import AscendW8A8MXFP8DynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.w4a4.w4a4_mxfp4 import AscendW4A4MXFP4DynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.w4a8.w4a8 import AscendW4A8DynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4 import AscendW4A8MXFPDynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.w8a8.w8a8_dynamic import AscendW8A8DynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.w8a8.w8a8_mxfp8 import AscendW8A8MXFP8DynamicFusedMoEMethod
+from vllm_ascend.quantization.methods.wna16.w4a16 import AscendW4A16FusedMoEMethod
+from vllm_ascend.quantization.methods.wna16.w4a16_mxfp4 import AscendW4A16MXFP4FusedMoEMethod
 
 
 def test_get_moe_num_logical_experts_uses_vllm_config_field():
