+++ tests/ut/quantization/methods/test_w4a8.py
@@ -5,8 +5,8 @@ import torch
 
 from tests.ut.base import TestBase
 from tests.ut.quantization.conftest_quantization import identity
-from vllm_ascend.quantization.methods.w4a8 import AscendW4A8DynamicFusedMoEMethod
-from vllm_ascend.utils import COMPRESSED_TENSORS_METHOD
+from vllm_ascend.quantization.methods.w4a8.w4a8 import AscendW4A8DynamicFusedMoEMethod
+from vllm_ascend.utils import ASCEND_QUANTIZATION_METHOD, COMPRESSED_TENSORS_METHOD
 
 
 class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
@@ -14,9 +14,9 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
     input_size = 16
     output_size = 56
 
-    @patch("vllm_ascend.quantization.methods.w4a8.get_ascend_config")
-    @patch("vllm_ascend.quantization.methods.w4a8.get_current_vllm_config")
-    @patch("vllm_ascend.quantization.methods.w4a8.get_mc2_group")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_ascend_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_mc2_group")
     @patch("torch.distributed.get_rank", return_value=0)
     def setUp(self, mock_get_rank, mock_get_mc2_group, get_current_vllm_config, mock_get_ascend_config):
         # Mock ascend config
@@ -25,7 +25,7 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
         mock_get_ascend_config.return_value = mock_ascend_config
 
         mock_vllm_config = Mock()
-        mock_vllm_config.quant_config = Mock(quant_description={"group_size": 0})
+        mock_vllm_config.quant_config = Mock(quant_description={"group_size": 0}, get_name=ASCEND_QUANTIZATION_METHOD)
         mock_vllm_config.parallel_config = Mock(enable_expert_parallel=True, enable_eplb=False)
         mock_vllm_config.use_v2_model_runner = False
         mock_vllm_config.scheduler_config = Mock(
@@ -34,15 +34,23 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
         get_current_vllm_config.return_value = mock_vllm_config
         self.quant_method = AscendW4A8DynamicFusedMoEMethod()
 
-    def test_init_rejects_per_group_quantization(self):
-        with patch("vllm_ascend.quantization.methods.w4a8.get_current_vllm_config") as mock_config:
-            mock_vllm_config = Mock()
-            mock_vllm_config.quant_config = Mock(quant_description={"group_size": 256})
-            mock_vllm_config.parallel_config = Mock(enable_expert_parallel=True)
-            mock_vllm_config.use_v2_model_runner = False
-            mock_config.return_value = mock_vllm_config
-            with self.assertRaisesRegex(ValueError, "no longer supported"):
-                AscendW4A8DynamicFusedMoEMethod()
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_tensor_model_parallel_world_size")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_current_vllm_config")
+    def test_init_raise_error(self, get_current_vllm_config, mock_tp_size):
+        # test init rejects per-group quantization
+        mock_vllm_config = Mock()
+        mock_vllm_config.quant_config = Mock(quant_description={"group_size": 256})
+        mock_vllm_config.parallel_config = Mock(enable_expert_parallel=True)
+        mock_vllm_config.use_v2_model_runner = False
+        get_current_vllm_config.return_value = mock_vllm_config
+        with self.assertRaisesRegex(ValueError, "no longer supported"):
+            AscendW4A8DynamicFusedMoEMethod()
+
+        # test init rejects tp>16 when use modelslim weights
+        mock_vllm_config.parallel_config = Mock(enable_expert_parallel=False)
+        mock_tp_size.return_value = 32
+        with self.assertRaisesRegex(ValueError, "tp>16"):
+            AscendW4A8DynamicFusedMoEMethod()
 
     def test_get_weight(self):
         param_dict = self.quant_method.get_weight(self.experts, self.input_size, self.output_size, torch.bfloat16)
@@ -175,8 +183,8 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
         ):
             self.quant_method.get_eplb_weight_views(layer)
 
-    @patch("vllm_ascend.quantization.methods.w4a8.get_ascend_config")
-    @patch("vllm_ascend.quantization.methods.w4a8.maybe_trans_nz")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_ascend_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.maybe_trans_nz")
     @patch("torch.Tensor.npu", new=lambda self: self, create=True)
     def test_process_weights_after_loading(self, mock_maybe_trans_nz, mock_get_ascend_config):
         mock_maybe_trans_nz.side_effect = identity
@@ -198,8 +206,8 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
         self.assertIs(weight_views[0], list_layer.w13_weight_list)
         self.assertIs(weight_views[-1], list_layer.w2_scale_bias_list)
 
-    @patch("vllm_ascend.quantization.methods.w4a8.get_ascend_config")
-    @patch("vllm_ascend.quantization.methods.w4a8.maybe_trans_nz")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.get_ascend_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.maybe_trans_nz")
     @patch("torch.Tensor.npu", new=lambda self: self, create=True)
     def test_process_weights_after_loading_compressed_tensors(self, mock_maybe_trans_nz, mock_get_ascend_config):
         mock_maybe_trans_nz.side_effect = identity
@@ -227,8 +235,8 @@ class TestAscendW4A8DynamicFusedMoEMethod(TestBase):
         with self.assertRaisesRegex(AssertionError, re.escape(expected_message)):
             self.quant_method._pack_to_int32(weight)
 
-    @patch("vllm_ascend.quantization.methods.w4a8._EXTRA_CTX")
-    @patch("vllm_ascend.quantization.methods.w4a8.build_fused_experts_input")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8._EXTRA_CTX")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8.build_fused_experts_input")
     def test_apply_comprehensive(self, mock_build_input, mock_ctx):
         tokens = 4
         num_experts = self.experts
