+++ tests/ut/quantization/methods/test_w4a8_mxfp4.py
@@ -5,14 +5,14 @@ import torch.nn as nn
 
 from tests.ut.base import TestBase
 from tests.ut.quantization.conftest_quantization import create_mock_ascend_config, create_mock_vllm_config
-from vllm_ascend.quantization.methods.w4a8_mxfp4 import (
+from vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4 import (
     AscendW4A8MXFPDynamicFusedMoEMethod,
     AscendW4A8MXFPDynamicLinearMethod,
 )
 
 
 class TestAscendW4A8MXFP4LinearMethod(TestBase):
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.get_current_vllm_config")
     def setUp(self, mock_vllm):
         mock_vllm.return_value = create_mock_vllm_config()
         self.scheme = AscendW4A8MXFPDynamicLinearMethod()
@@ -31,7 +31,7 @@ class TestAscendW4A8MXFP4LinearMethod(TestBase):
             self.assertEqual(result["weight_scale"].shape, (128, 256 // gs))
             self.assertEqual(result["weight_scale"].dtype, torch.uint8)
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.torch_npu")
     def test_process_weights_after_loading_transposes(self, mock_npu):
         # npu_format_cast returns the input tensor (mocked as identity)
         mock_npu.npu_format_cast.side_effect = lambda x, *a, **kw: x
@@ -42,7 +42,7 @@ class TestAscendW4A8MXFP4LinearMethod(TestBase):
         self.assertEqual(layer.weight.shape, (128, 128))
         self.assertEqual(layer.weight_scale.shape, (4, 128, 2))
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.torch_npu")
     def test_apply_with_prequantized_input(self, mock_npu):
         mock_npu.npu_quant_matmul.return_value = torch.randn(32, 128)
         layer = MagicMock()
@@ -55,7 +55,7 @@ class TestAscendW4A8MXFP4LinearMethod(TestBase):
             output = self.scheme.apply(layer, x)
         self.assertEqual(output.shape[0], 32)
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.torch_npu")
     def test_apply_with_bf16_input(self, mock_npu):
         mock_npu.npu_dynamic_mx_quant.return_value = (
             torch.randint(0, 255, (32, 128), dtype=torch.uint8),
@@ -76,13 +76,11 @@ class TestAscendW4A8MXFP4MoEMethod(TestBase):
     hidden_size = 128
     intermediate_size = 256
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.get_ep_group")
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.get_current_vllm_config")
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.get_ascend_config")
-    def setUp(self, mock_ascend, mock_vllm, mock_ep_group):
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.get_current_vllm_config")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.get_ascend_config")
+    def setUp(self, mock_ascend, mock_vllm):
         mock_vllm.return_value = create_mock_vllm_config()
         mock_ascend.return_value = create_mock_ascend_config()
-        mock_ep_group.return_value = Mock()
         self.scheme = AscendW4A8MXFPDynamicFusedMoEMethod()
 
     def test_get_weight_static_method(self):
@@ -105,7 +103,7 @@ class TestAscendW4A8MXFP4MoEMethod(TestBase):
             self.assertEqual(result["w13_weight_scale"].dtype, torch.uint8)
             self.assertEqual(result["w2_weight_scale"].dtype, torch.uint8)
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.torch_npu")
     def test_process_weights_transposes_weights(self, mock_npu):
         # npu_format_cast returns the input tensor (mocked as identity)
         mock_npu.npu_format_cast.side_effect = lambda x, *a, **kw: x
@@ -122,8 +120,8 @@ class TestAscendW4A8MXFP4MoEMethod(TestBase):
         self.assertEqual(layer.w13_weight_scale.shape, (8, 2, 256, 2))
         self.assertEqual(layer.w2_weight_scale.shape, (8, 4, 128, 2))
 
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4.torch_npu")
-    @patch("vllm_ascend.quantization.methods.w4a8_mxfp4._EXTRA_CTX")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4.torch_npu")
+    @patch("vllm_ascend.quantization.methods.w4a8.w4a8_mxfp4._EXTRA_CTX")
     def test_apply_full_params(self, mock_ctx, mock_npu):
         tokens = 4
         layer = nn.Module()
