+++ tests/ut/quantization/test_modelslim_config.py
@@ -15,7 +15,7 @@ from vllm.model_executor.models.utils import WeightsMapper
 from tests.ut.base import TestBase
 from vllm_ascend.models.llama_eagle3 import get_rotation_path
 from vllm_ascend.ops.linear import AscendUnquantizedLinearMethod
-from vllm_ascend.quantization.modelslim_config import (
+from vllm_ascend.quantization.configs.modelslim_config import (
     MODELSLIM_CONFIG_FILENAME,
     AscendModelSlimConfig,
     _make_modelslim_moe_weight_loader,
@@ -93,7 +93,9 @@ class TestAscendModelSlimConfig(TestBase):
         linear_layer = MagicMock(spec=LinearBase)
         # Test skipped layer
         with (
-            patch("vllm_ascend.quantization.modelslim_config.get_current_vllm_config", return_value=mock_config),
+            patch(
+                "vllm_ascend.quantization.configs.modelslim_config.get_current_vllm_config", return_value=mock_config
+            ),
             patch.object(self.ascend_config, "is_layer_skipped_ascend", return_value=True),
         ):
             method = self.ascend_config.get_quant_method(linear_layer, ".attn")
@@ -103,8 +105,12 @@ class TestAscendModelSlimConfig(TestBase):
         mock_scheme = MagicMock()
         with (
             patch.object(self.ascend_config, "is_layer_skipped_ascend", return_value=False),
-            patch("vllm_ascend.quantization.modelslim_config.get_current_vllm_config", return_value=mock_config),
-            patch("vllm_ascend.quantization.modelslim_config.create_scheme_for_layer", return_value=mock_scheme),
+            patch(
+                "vllm_ascend.quantization.configs.modelslim_config.get_current_vllm_config", return_value=mock_config
+            ),
+            patch(
+                "vllm_ascend.quantization.configs.modelslim_config.create_scheme_for_layer", return_value=mock_scheme
+            ),
             patch(
                 "vllm_ascend.quantization.method_adapters.AscendLinearMethod", return_value=MagicMock()
             ) as mock_ascend_linear,
@@ -119,8 +125,12 @@ class TestAscendModelSlimConfig(TestBase):
         mock_config.model_config.hf_config.model_type = None
         mock_scheme = MagicMock()
         with (
-            patch("vllm_ascend.quantization.modelslim_config.get_current_vllm_config", return_value=mock_config),
-            patch("vllm_ascend.quantization.modelslim_config.create_scheme_for_layer", return_value=mock_scheme),
+            patch(
+                "vllm_ascend.quantization.configs.modelslim_config.get_current_vllm_config", return_value=mock_config
+            ),
+            patch(
+                "vllm_ascend.quantization.configs.modelslim_config.create_scheme_for_layer", return_value=mock_scheme
+            ),
             patch(
                 "vllm_ascend.quantization.method_adapters.AscendKVCacheMethod", return_value=MagicMock()
             ) as mock_ascend_kvcache,
@@ -214,9 +224,12 @@ class TestAscendModelSlimConfig(TestBase):
         mock_vllm_config_for_kv_c8.kv_transfer_config = None
 
         with (
-            patch("vllm_ascend.quantization.modelslim_config.get_current_vllm_config", return_value=mock_vllm_config),
             patch(
-                "vllm_ascend.quantization.methods.kv_c8.get_current_vllm_config",
+                "vllm_ascend.quantization.configs.modelslim_config.get_current_vllm_config",
+                return_value=mock_vllm_config,
+            ),
+            patch(
+                "vllm_ascend.quantization.methods.kv_cache.kv_c8.get_current_vllm_config",
                 return_value=mock_vllm_config_for_kv_c8,
             ),
             patch(
@@ -226,7 +239,7 @@ class TestAscendModelSlimConfig(TestBase):
             method = c8_config.get_quant_method(attention_layer, "model.layers.0.self_attn.attn")
             self.assertIs(method, mock_kvcache.return_value)
             args, _ = mock_kvcache.call_args
-            from vllm_ascend.quantization.methods.kv_c8 import AscendC8KVCacheAttentionMethod
+            from vllm_ascend.quantization.methods.kv_cache.kv_c8 import AscendC8KVCacheAttentionMethod
 
             self.assertIsInstance(args[0], AscendC8KVCacheAttentionMethod)
 
@@ -419,7 +432,7 @@ class TestGetCacheScaleMapper(TestBase):
         config = AscendModelSlimConfig({"kv_cache_type": "C8"})
         mapper = config.get_cache_scale_mapper()
         self.assertIsNotNone(mapper)
-        # C8 mappings: k_proj → attn
+        # C8 mappings: k_proj -> attn
         self.assertEqual(
             mapper._map_name("model.layers.0.k_proj.kv_cache_scale"),
             "model.layers.0.attn.k_cache_scale",
@@ -733,7 +746,7 @@ class TestGetKvQuantDtype(TestBase):
 
 
 class TestGetKvQuantSplitFactor(TestBase):
-    @patch("vllm_ascend.quantization.modelslim_config.calc_split_factor")
+    @patch("vllm_ascend.quantization.configs.modelslim_config.calc_split_factor")
     def test_enable_fa_quant_true(self, mock_calc_split_factor):
         mock_calc_split_factor.return_value = 2.0
         config = AscendModelSlimConfig(
@@ -748,7 +761,7 @@ class TestGetKvQuantSplitFactor(TestBase):
         self.assertEqual(result, 2.0)
         mock_calc_split_factor.assert_called_once_with([64, 128])
 
-    @patch("vllm_ascend.quantization.modelslim_config.calc_split_factor")
+    @patch("vllm_ascend.quantization.configs.modelslim_config.calc_split_factor")
     def test_enable_fa_quant_false(self, mock_calc_split_factor):
         mock_calc_split_factor.return_value = 1.0
         config = AscendModelSlimConfig({})
