+++ vllm_ascend/quantization/methods/w4a8/__init__.py
@@ -1,0 +1,31 @@
+#
+# Copyright (c) 2026 Huawei Technologies Co., Ltd. All Rights Reserved.
+# This file is a part of the vllm-ascend project.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""W4A8 quantization methods for Ascend NPU."""
+
+from .w4a8 import AscendW4A8DynamicFusedMoEMethod
+from .w4a8_mxfp4 import (
+    AscendW4A8MXFPDSDynamicFusedMoEMethod,
+    AscendW4A8MXFPDynamicFusedMoEMethod,
+    AscendW4A8MXFPDynamicLinearMethod,
+)
+
+__all__ = [
+    "AscendW4A8DynamicFusedMoEMethod",
+    "AscendW4A8MXFPDSDynamicFusedMoEMethod",
+    "AscendW4A8MXFPDynamicFusedMoEMethod",
+    "AscendW4A8MXFPDynamicLinearMethod",
+]
