+++ vllm_ascend/quantization/methods/w4a8/w4a8.py
@@ -29,8 +29,8 @@ from vllm_ascend.ops.fused_moe.dataclass.fused_experts import build_fused_expert
 from vllm_ascend.ops.fused_moe.routed_experts import AscendRoutedExperts  # noqa: F401
 from vllm_ascend.utils import ASCEND_QUANTIZATION_METHOD, COMPRESSED_TENSORS_METHOD, maybe_trans_nz
 
-from .base import AscendMoEScheme, QuantType
-from .registry import register_scheme
+from ..base import AscendMoEScheme, QuantType
+from ..registry import register_scheme
 
 
 @register_scheme("W4A8_DYNAMIC", "moe")
@@ -38,10 +38,7 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
     """FusedMoE method for Ascend W4A8_DYNAMIC.
 
     This method supports two MoE weight formats: one generated by
-    msModelSlim and one generated by LLM-Compressor. The LLM-Compressor
-    path is selected when ``ascend_quant_method`` in
-    ``quant_description`` is ``COMPRESSED_TENSORS_METHOD``. Otherwise,
-    the msModelSlim path is used. Only per-channel weight quantization
+    msModelSlim and one generated by LLM-Compressor. Only per-channel weight quantization
     is supported; per-group quantization is rejected at init time.
 
     The names below use ``L`` for the layer index, ``E`` for the expert
@@ -119,7 +116,7 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
                 "switch to Per‑Channel quantized weights."
             )
 
-        self.quant_method = vllm_config.quant_config.quant_description.get("ascend_quant_method", "")
+        self.quant_method = vllm_config.quant_config.get_name()
         self.tp_size = (
             1 if vllm_config.parallel_config.enable_expert_parallel else get_tensor_model_parallel_world_size()
         )
@@ -405,7 +402,7 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
         # To use torch_npu.npu_grouped_matmul, keep w2_weigh_scale unsqueezed
         layer.w13_weight_scale.data = self.maybe_squeeze_per_channel_weight_scale(layer.w13_weight_scale.data)
 
-        # Packs 2 int4 into 1 int8 on-the-fly to mirror the modelslim new_quant_version path
+        # Packs 2 int4 into 1 int8 on-the-fly to mirror the modelslim path
         layer.w13_weight.data = self._pack_int4_to_int8(layer.w13_weight.data)
         layer.w2_weight.data = self._pack_int4_to_int8(layer.w2_weight.data)
