+++ vllm_ascend/quantization/methods/w8a8/w8a8_pdmix.py
@@ -27,8 +27,8 @@ from typing import Any
 import torch
 from vllm.config import get_current_vllm_config
 
-from .base import AscendLinearScheme
-from .registry import register_scheme
+from ..base import AscendLinearScheme
+from ..registry import register_scheme
 from .w8a8_dynamic import AscendW8A8DynamicLinearMethod
 from .w8a8_static import AscendW8A8LinearMethod
