+++ vllm_ascend/quantization/methods/wna16/__init__.py
@@ -1,0 +1,27 @@
+#
+# Copyright (c) 2026 Huawei Technologies Co., Ltd. All Rights Reserved.
+# This file is a part of the vllm-ascend project.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""WnA16 quantization methods for Ascend NPU (W4A16, W8A16)."""
+
+from .w4a16 import AscendW4A16FusedMoEMethod
+from .w4a16_mxfp4 import AscendW4A16MXFP4FusedMoEMethod
+from .w8a16 import AscendW8A16LinearMethod
+
+__all__ = [
+    "AscendW4A16FusedMoEMethod",
+    "AscendW4A16MXFP4FusedMoEMethod",
+    "AscendW8A16LinearMethod",
+]
