+++ vllm_ascend/quantization/methods/wna16/w4a16.py
@@ -27,8 +27,8 @@ from vllm_ascend.ascend_forward_context import _EXTRA_CTX
 from vllm_ascend.ops.fused_moe.dataclass.fused_experts import build_fused_experts_input
 from vllm_ascend.ops.fused_moe.routed_experts import AscendRoutedExperts  # noqa: F401
 
-from .base import AscendMoEScheme, QuantType
-from .registry import register_scheme
+from ..base import AscendMoEScheme, QuantType
+from ..registry import register_scheme
 
 
 def unpack_from_int32(
