+++ vllm_ascend/quantization/methods/wna16/w4a16_mxfp4.py
@@ -20,16 +20,15 @@ from typing import Any
 
 import torch
 import torch_npu
-from vllm.config import CompilationMode, get_current_vllm_config
-from vllm.distributed import get_ep_group
+from vllm.config import get_current_vllm_config
 
 from vllm_ascend.ascend_config import get_ascend_config
 from vllm_ascend.ascend_forward_context import _EXTRA_CTX
 from vllm_ascend.ops.fused_moe.dataclass.fused_experts import build_fused_experts_input
 from vllm_ascend.ops.fused_moe.routed_experts import AscendRoutedExperts  # noqa: F401
 
-from .base import AscendMoEScheme, QuantType
-from .registry import register_scheme
+from ..base import AscendMoEScheme, QuantType
+from ..registry import register_scheme
 
 
 # Unpack the weights to FP4 and return them in float32 format
@@ -55,15 +54,9 @@ class AscendW4A16MXFP4FusedMoEMethod(AscendMoEScheme):
     quant_type: QuantType = QuantType.W4A16MXFP
 
     def __init__(self) -> None:
-        self.ep_group = get_ep_group()
-
         vllm_config = get_current_vllm_config()
         self.group_size = vllm_config.quant_config.quant_description.get("group_size", 32)
         ascend_config = get_ascend_config()
-        self.use_aclgraph = (
-            vllm_config.compilation_config.mode == CompilationMode.VLLM_COMPILE
-            and not vllm_config.model_config.enforce_eager
-        )
         self.dynamic_eplb = False if vllm_config.use_v2_model_runner else ascend_config.eplb_config.dynamic_eplb
 
     def get_weight(
